+++ src/Components/QuizQuestion/QuizQuestions.types.ts
@@ -1,0 +1,5 @@
+import { QUESTION } from "../../data/data.types";
+
+export type QUESTIONS = {
+  questions: QUESTION[];
+};
