+++ src/context/quizdata-context.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext } from "react";
 import { data } from "../data/data";
-import { QUIZCATEGORIES } from "../data/data";
+import { QUIZCATEGORIES } from "../data/data.types";
 export const DataContext = createContext<QUIZCATEGORIES | null>(null);
 
 export function DataProvider({ children }: React.PropsWithChildren<{}>) {
